AI_Employee_Vault/Automation/watchers/requirements.txt
# AI Employee Watchers - Python Dependencies

# Base dependencies
python-dateutil>=2.8.2

# Gmail Watcher
google-auth>=2.16.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.70.0

# WhatsApp & LinkedIn Watchers
playwright>=1.40.0

# Optional: Better async support
asyncio>=3.4.3
